Split mongodb connection and schema hooks into module
